+++ src/cmd/bins.rs
@@ -5,7 +5,7 @@ use rayon::slice::ParallelSliceMut;
 
 use crate::config::{Config, Delimiter};
 use crate::scales::LinearScale;
-use crate::select::SelectColumns;
+use crate::select::SelectedColumns;
 use crate::util;
 use crate::CliResult;
 
@@ -49,7 +49,7 @@ Common options:
 #[derive(Deserialize)]
 struct Args {
     arg_input: Option<String>,
-    flag_select: SelectColumns,
+    flag_select: SelectedColumns,
     flag_no_headers: bool,
     flag_delimiter: Option<Delimiter>,
     flag_output: Option<String>,