+++ src/json.rs
@@ -1,13 +1,14 @@
 use std::borrow::Cow;
 use std::cell::RefCell;
 use std::collections::{btree_map::Entry as BTreeMapEntry, BTreeMap};
-use std::io::Read;
+use std::io::{Read, Write};
 use std::num::NonZeroUsize;
 use std::rc::Rc;
 
 use csv::StringRecord;
 use serde::ser::{Serialize, SerializeMap, Serializer};
 use serde_json::{json, Value};
+use simd_csv::ByteRecord;
 
 use crate::select::Selection;
 
@@ -168,12 +169,12 @@ fn traverse_to_build_stack(value: &Value, stack: &mut JSONTraversalStack, depth:
     };
 }
 
-fn headers_from_stack(stack: &JSONTraversalStack) -> csv::StringRecord {
-    let mut record = csv::StringRecord::new();
+fn headers_from_stack(stack: &JSONTraversalStack) -> ByteRecord {
+    let mut record = ByteRecord::new();
 
     // Single scalar early return
     if stack.len() == 1 {
-        record.push_field("value");
+        record.push_field(b"value");
         return record;
     }
 
@@ -185,7 +186,7 @@ fn headers_from_stack(stack: &JSONTraversalStack) -> csv::StringRecord {
                 path.push(key.as_str());
             }
             JSONTraversalState::Emit => {
-                record.push_field(&path.join("."));
+                record.push_field(path.join(".").as_bytes());
             }
             JSONTraversalState::Pop(_) => {
                 path.pop();
@@ -247,18 +248,20 @@ where
     }
 }
 
-fn serialize_json_value_to_csv_field(value: &Value) -> Cow<str> {
+#[inline]
+fn serialize_json_value_to_csv_field(value: &Value) -> Cow<[u8]> {
     match value {
-        Value::Null => Cow::Borrowed(""),
-        Value::Bool(b) => Cow::Borrowed(if *b { "true" } else { "false" }),
-        Value::String(s) => Cow::Borrowed(s.as_str()),
-        Value::Number(n) => Cow::Owned(n.to_string()),
-        Value::Array(l) => Cow::Owned(serde_json::to_string(l).unwrap()),
-        Value::Object(o) => Cow::Owned(serde_json::to_string(o).unwrap()),
+        Value::Null => Cow::Borrowed(b""),
+        Value::Bool(b) => Cow::Borrowed(if *b { b"true" } else { b"false" }),
+        Value::String(s) => Cow::Borrowed(s.as_bytes()),
+        Value::Number(n) => Cow::Owned(n.to_string().into_bytes()),
+        Value::Array(l) => Cow::Owned(serde_json::to_vec(l).unwrap()),
+        Value::Object(o) => Cow::Owned(serde_json::to_vec(o).unwrap()),
     }
 }
 
-fn fill_record(value: &Value, record: &mut StringRecord, stack: &JSONTraversalStack) {
+#[inline]
+fn fill_record(value: &Value, record: &mut ByteRecord, stack: &JSONTraversalStack) {
     record.clear();
 
     traverse_with_stack(value, stack, |v| {
@@ -280,63 +283,65 @@ fn merge(a: &mut Value, b: &Value) {
     *a = b.clone();
 }
 
-pub fn for_each_json_value_as_csv_record<I, F, E>(
-    values: I,
-    sample_size: NonZeroUsize,
-    mut callback: F,
-) -> Result<(), E>
-where
-    I: Iterator<Item = Result<Value, E>>,
-    F: FnMut(&StringRecord) -> Result<(), E>,
-{
-    let mut merged_value_from_sample = Value::Null;
-    let mut sampled_records: Vec<Value> = Vec::new();
-    let mut headers_emitted: bool = false;
-    let mut output_record = StringRecord::new();
-    let mut stack = JSONTraversalStack::new();
+pub struct JSONTabularizer<W: Write> {
+    writer: simd_csv::Writer<W>,
+    harmonized_value: Value,
+    sample_size: usize,
+    sample: Vec<Value>,
+    flushed: bool,
+    output_record: ByteRecord,
+    stack: JSONTraversalStack,
+}
 
-    let sample_size: usize = sample_size.into();
-
-    for (i, result) in values.enumerate() {
-        let value = result?;
-
-        // Reading sample
-        if i < sample_size {
-            merge(&mut merged_value_from_sample, &value);
-            sampled_records.push(value);
-            continue;
-        }
-
-        // Emitting headers
-        if !headers_emitted {
-            traverse_to_build_stack(&merged_value_from_sample, &mut stack, 0);
-            callback(&headers_from_stack(&stack))?;
-
-            for sample in sampled_records.iter() {
-                fill_record(sample, &mut output_record, &stack);
-                callback(&output_record)?;
-            }
-
-            headers_emitted = true;
-            sampled_records.clear();
-        }
-
-        fill_record(&value, &mut output_record, &stack);
-        callback(&output_record)?;
-    }
-
-    // Sample was larger than the file
-    if !sampled_records.is_empty() {
-        traverse_to_build_stack(&merged_value_from_sample, &mut stack, 0);
-        callback(&headers_from_stack(&stack))?;
-
-        for sample in sampled_records.iter() {
-            fill_record(sample, &mut output_record, &stack);
-            callback(&output_record)?;
+impl<W: Write> JSONTabularizer<W> {
+    pub fn from_writer(writer: simd_csv::Writer<W>, sample_size: NonZeroUsize) -> Self {
+        Self {
+            writer,
+            harmonized_value: Value::Null,
+            sample_size: sample_size.get(),
+            sample: Vec::with_capacity(sample_size.get()),
+            flushed: false,
+            output_record: ByteRecord::new(),
+            stack: JSONTraversalStack::new(),
         }
     }
 
-    Ok(())
+    pub fn flush(&mut self) -> simd_csv::Result<()> {
+        if self.flushed {
+            return Ok(());
+        }
+
+        traverse_to_build_stack(&self.harmonized_value, &mut self.stack, 0);
+        self.writer
+            .write_byte_record(&headers_from_stack(&self.stack))?;
+
+        for value in self.sample.iter() {
+            fill_record(value, &mut self.output_record, &self.stack);
+            self.writer.write_byte_record(&self.output_record)?;
+        }
+
+        self.writer.flush()?;
+
+        self.flushed = true;
+
+        Ok(())
+    }
+
+    pub fn process(&mut self, value: Value) -> simd_csv::Result<()> {
+        // Sampling
+        if self.sample.len() < self.sample_size {
+            merge(&mut self.harmonized_value, &value);
+            self.sample.push(value);
+            return Ok(());
+        }
+
+        self.flush()?;
+
+        fill_record(&value, &mut self.output_record, &self.stack);
+        self.writer.write_byte_record(&self.output_record)?;
+
+        Ok(())
+    }
 }
 
 const JSON_MAX_SAFE_INTEGER: i64 = 9007199254740991;